Editor: Graph Node changes it's position on click in graph editor #862

+++ engine/includes/editor/viewport/viewport.h
@@ -33,7 +33,7 @@ public:
 
     PipelineContext *pipelineContext() const;
 
-    float gridCell();
+    int gridCell();
 
     bool isGamePaused() const;
     void setGamePaused(bool pause);
